+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.13.5] 2025-05-03
+### Added
+- Added a new "On This Day" section to the stats to see what songs you had on repeat in past years on the same date.
+
+### Changed
+- Cleared the default directory in `config.py` since the `sesh` folder is gone.
+
 ## [1.13.4] 2025-05-03
 ### Changed
 - Updated GUI text to be more helpful.